refact: category icon을 사용하는 곳을 한 곳으로 일원화합니다.
